feat!: Data provider and data is now separated
